refactor: introduce Job and Step structures for GitHub workflow management
