+++ crates/gh-workflow/src/lib.rs
@@ -3,6 +3,8 @@ extern crate proc_macro;
 mod cargo;
 pub mod error;
 mod event;
+mod job;
+mod step;
 
 pub mod ctx;
 pub mod generate;
@@ -13,7 +15,9 @@ pub(crate) mod workflow;
 
 pub use cargo::*;
 pub use event::*;
+pub use job::*;
 pub use rust_flag::*;
+pub use step::*;
 pub use workflow::*;
 
 pub(crate) fn is_default<T: Default + PartialEq>(value: &T) -> bool {
